pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.2"
python-telegram-bot = "^22.5.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
minio = "^7.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

### Ruff ###

[tool.ruff]
target-version = "py312"
line-length = 120  # 100 или 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "E501",  # Line too long
  "D",  # pydocstyle
  "S",  # flake8-bandit
  "W",  # pycodestyle warnings (покрываются форматированием)
  "EM", "TRY003",  # Передача строк в конструктор исключений
  "TD", "FIX",  # flake8-todos, flake8-fixme
  "TCH",  # flake8-type-checking
  "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
  "RUF001", "RUF002", "RUF003",  # ambiguous character warnings (Не работают с кириллицей)
  "COM812", "ISC001",  # following rules may cause conflicts when used with the formatter
  "PGH003",  # Use specific rule codes when ignoring type issues
  "BLE001",  # Do not catch blind exception: `Exception`
  "ERA",  # Found commented-out code
  "PLR0913",  # Too many arguments in function definition
  "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.per-file-ignores]
"alembic/**/*.py" = [
  "INP", "E402"
]
"*.ipynb" = [
  "PLE1142",  # Checks for uses of await outside of async functions
  "T201",  # Checks for print statements
]

### Pyrgiht ###

[tool.pyright]
exclude = [".venv/", "**/__pycache__", ".*cache*/", "__*/", "alembic/versions/"]
venvPath = "."
venv = ".venv"

# typeCheckingMode = "strict"  # Строгий режим не нужно соблюдать постоянно, но полезно иногда включать и проверять

reportUnnecessaryTypeIgnoreComment = true
reportMissingTypeArgument = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true